Create Program_to_Print_Job_Sequencing
Given an array of jobs where every job has a deadline and profit if the job is finished before the deadline. 
It is also given that every job takes a single unit of time, so the minimum possible deadline for any job is 1.
Maximize the total profit if only one job can be scheduled at a time.